playwright-python/requirements.txt
pytest
pytest-playwright
pytest-base-url